Fix eclipse warnings
A regular sweep aimed at cleaning the code up.

+++ openflowplugin-impl/src/test/java/org/opendaylight/openflowplugin/impl/connection/OutboundQueueProviderImplTest.java
@@ -1,13 +1,12 @@
 package org.opendaylight.openflowplugin.impl.connection;
 
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 import junit.framework.TestCase;
 import org.junit.Test;
 import org.opendaylight.openflowjava.protocol.api.connection.OutboundQueue;
 import org.opendaylight.openflowplugin.api.OFConstants;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.protocol.rev130731.BarrierInput;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
 public class OutboundQueueProviderImplTest extends TestCase {
 
@@ -34,7 +33,7 @@ public class OutboundQueueProviderImplTest extends TestCase {
     public void testCreateBarrierRequest() {
         final BarrierInput barrierRequest = outboundQueueProvider.createBarrierRequest(DUMMY_XID);
         assertNotNull(barrierRequest);;
-        assertEquals((short)OFConstants.OFP_VERSION_1_3, (short)barrierRequest.getVersion());
+        assertEquals(OFConstants.OFP_VERSION_1_3, (short)barrierRequest.getVersion());
         assertEquals(DUMMY_XID, barrierRequest.getXid());
     }
 }
